功能基本完成

+++ src/main/java/cloud/base/controller/ResourceController.java
@@ -1,10 +1,12 @@
 package cloud.base.controller;
 
+import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -16,6 +18,7 @@ import cloud.base.model.SysResource;
 import cloud.base.model.VO.PageData;
 import cloud.base.model.VO.PageModel;
 import cloud.base.service.ISysResourceService;
+import cloud.base.util.DictionaryUtil;
 
 @Controller
 @RequestMapping("/resource")
@@ -41,7 +44,17 @@ public class ResourceController {
 		
 		Map cond = pageModel.initPageModel();
 		sysresourceservice.loadPageModel(pageModel);
+		List<SysResource> list = pageModel.getPageData().getRows();
+		for (SysResource re : list) {
+			re.setResourcetypename(DictionaryUtil.getNameByTypeAndType("resourceType", re.getResourcetype()));
+		}
 		return pageModel.getPageData();
+	}
+	
+	@RequestMapping("/roleselectresource/{rolecode}")
+	public String roleselectresource(ModelMap modelMap,@PathVariable("rolecode") String rolecode){
+		modelMap.addAttribute("rolecode",rolecode);
+		return "/resource/roleselectresource";
 	}
 	
 	@RequestMapping("/add/{groupcode}")
@@ -78,8 +91,23 @@ public class ResourceController {
 	 */
 	@RequestMapping("/delete")
 	public @ResponseBody String delete(String ids){
-		sysresourceservice.deleteSysResource(ids.split(","));
+		if(!StringUtils.isEmpty(ids)){
+			String[] idsArray = ids.split(",");
+			for (String id : idsArray) {
+				//如果资源被引用，则不允许删除
+			}
+			sysresourceservice.deleteSysResource(ids.split(","));
+		}
 		return null;
+	}
+	/**
+	 * @param userid
+	 * 		根据角色id得到对应的资源
+	 * @return
+	 */
+	@RequestMapping("/getResourcesByRoleCode/{rolecode}")
+	public @ResponseBody List getResourcesByRoleCode(@PathVariable("rolecode") String rolecode){
+		return sysresourceservice.findAllResourceByRoleCode(rolecode);
 	}
 }
