+++ src/main/java/cloud/base/controller/ResourceGroupController.java
@@ -1,6 +1,5 @@
 package cloud.base.controller;
 
-import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -86,7 +85,16 @@ public class ResourceGroupController {
 	 */
 	@RequestMapping("/delete")
 	public @ResponseBody String delete(String groupids){
-		resourcegroupservice.delete(groupids.split(","));
+		if(!StringUtils.isEmpty(groupids)){
+			String[] idsArray = groupids.split(",");
+			for (String id : idsArray) {
+				//如果有子节点，或者组里有角色记录，不允许删除
+				if(resourcegroupservice.groupIsHaschildren(id)>0){
+					
+				}
+			}
+			resourcegroupservice.delete(groupids.split(","));
+		}
 		return null;
 	}
 }
