+++ src/main/java/cloud/base/controller/UserGroupController.java
@@ -86,7 +86,17 @@ public class UserGroupController {
 	 */
 	@RequestMapping("/delete")
 	public @ResponseBody String delete(String groupids){
-		usergroupservice.delete(groupids.split(","));
+		if(!StringUtils.isEmpty(groupids)){
+			String[] idsArray = groupids.split(",");
+			for (String id : idsArray) {
+				//如果有子节点，或者组里有角色记录，不允许删除
+				if(usergroupservice.groupIsHaschildren(id)>0){
+					
+				}
+			}
+			
+			usergroupservice.delete(groupids.split(","));
+		}
 		return null;
 	}
 }
